requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arxiv==2.4.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
duckduckgo_search==8.1.1
emoji==2.15.0
fastapi==0.128.0
fastuuid==0.14.0
feedparser==6.0.12
filelock==3.20.3
filetype==1.2.0
fonttools==4.61.1
fpdf2==2.8.5
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
gpt-researcher==0.14.5
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
htmldocx==0.0.6
httpcore==1.0.9
httpx==0.28.1
httpx-aiohttp==0.1.12
httpx-sse==0.4.3
huggingface_hub==1.3.1
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
json5==0.13.0
json_repair==0.55.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==1.2.3
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-mcp-adapters==0.2.1
langchain-ollama==1.0.1
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langdetect==1.0.9
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-cli==0.4.11
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.2
litellm==1.80.16
loguru==0.7.3
lxml==6.0.2
Markdown==3.10
markdown2==2.5.4
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.25.0
md2pdf==2.1.0
mistune==3.2.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.2.6
olefile==0.47
ollama==0.6.1
openai==2.15.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
primp==0.15.0
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydyf==0.12.1
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.20
PyMuPDF==1.26.7
pypdf==6.6.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-frontmatter==1.1.0
python-iso639==2025.11.16
python-magic==0.4.27
python-multipart==0.0.21
python-oxmsg==0.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.7
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
sse-starlette==3.1.2
starlette==0.50.0
streamlit==1.52.2
tavily-python==0.7.17
tenacity==9.1.2
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.18.27
unstructured-client==0.42.6
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
watchdog==6.0.0
weasyprint==67.0
webencodings==0.5.1
websockets==16.0
wheel==0.45.1
win32_setctime==1.2.0
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zopfli==0.4.0
zstandard==0.25.0
